microservices/signature-service/requirements.txt
# BuffrSign-Starter: Signature Service Requirements
# Core signature envelope management and digital signature processing

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Supabase integration
supabase==2.3.0

# Database and async support
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2